Move themeable decorator and make Checklist always themeable
