+++ Checklist/index.jsx
@@ -1,5 +1,6 @@
 import React, { PropTypes } from 'react'
 import classNamesBind from 'classnames/bind'
+import themeable from '../decorators/themeable'
 import defaultStyles from './styles.scss'
 
 const baseClass = 'checklist'
@@ -9,7 +10,7 @@ const classes = {
   checkmark: `${baseClass}__checkmark`
 }
 
-export function Main ({ chromeless, className, children, customize, style, styles }) {
+function ChecklistMain ({ chromeless, className, children, customize, style, styles }) {
   const classNames = classNamesBind.bind({ ...defaultStyles, ...styles })
 
   const dynamicStyles = customize
@@ -18,21 +19,19 @@ export function Main ({ chromeless, className, children, customize, style, style
       borderColor: customize.borderColor
     } : {}
 
-  return (
-    <ul
-      style={{
-        ...dynamicStyles,
-        ...style
-      }}
-      className={classNames(baseClass, { chromeless }, className)}>
-      {children}
-    </ul>
-  )
+  return <ul
+    style={{
+      ...dynamicStyles,
+      ...style
+    }}
+    className={classNames(baseClass, { chromeless }, className)}>
+    {children}
+  </ul>
 }
 
-Main.displayName = 'Checklist.Main'
+ChecklistMain.displayName = 'Checklist.Main'
 
-Main.propTypes = {
+ChecklistMain.propTypes = {
   className: PropTypes.string,
   children: PropTypes.node,
   chromeless: PropTypes.bool,
@@ -43,7 +42,15 @@ Main.propTypes = {
   })
 }
 
-export function Item ({ className, children, customize, styles }) {
+export const Main = themeable(ChecklistMain, (customizations, props) => ({
+  customize: {
+    ...props.customize,
+    borderColor: customizations.color_border,
+    borderRadius: customizations.radius_border
+  }
+}))
+
+function ChecklistItem ({ className, children, customize, styles }) {
   const classNames = classNamesBind.bind({ ...defaultStyles, ...styles })
   const listItemDynamicStyles = customize
     ? { color: customize.textColor }
@@ -52,27 +59,25 @@ export function Item ({ className, children, customize, styles }) {
     ? { stroke: customize.strokeColor }
     : undefined
 
-  return (
-    <li
-      className={classNames(classes.item, className)}
-      style={listItemDynamicStyles}>
-      <svg
-        className={classNames(classes.checkmark)}
-        style={iconDynamicStyles}
-        viewBox='0 0 25 25'
-        aria-labelledby='Checkmark'
-        height='20px'
-        width='20px'>
-        <path d='M5 13.69l4.49 4.23L19.37 8'></path>
-      </svg>
-      {children}
-    </li>
-  )
+  return <li
+    className={classNames(classes.item, className)}
+    style={listItemDynamicStyles}>
+    <svg
+      className={classNames(classes.checkmark)}
+      style={iconDynamicStyles}
+      viewBox='0 0 25 25'
+      aria-labelledby='Checkmark'
+      height='20px'
+      width='20px'>
+      <path d='M5 13.69l4.49 4.23L19.37 8'></path>
+    </svg>
+    {children}
+  </li>
 }
 
-Item.displayName = 'Checklist.Item'
+ChecklistItem.displayName = 'Checklist.Item'
 
-Item.propTypes = {
+ChecklistItem.propTypes = {
   className: PropTypes.string,
   children: PropTypes.node,
   styles: PropTypes.object,
@@ -80,3 +85,11 @@ Item.propTypes = {
     strokeColor: PropTypes.string.isRequired
   })
 }
+
+export const Item = themeable(ChecklistItem, (customizations, props) => ({
+  customize: {
+    ...props.customize,
+    strokeColor: customizations.color_details,
+    textColor: customizations.color_text
+  }
+}))
