+++ themeable/Button/Primary.jsx
@@ -1,5 +1,5 @@
 import Primary from '../../Button/Primary'
-import themeable from '../../lib/decorators/themeable'
+import themeable from '../../decorators/themeable'
 
 const Component = themeable(Primary, (customizations, { customize }) => ({
   customize: {
