+++ themeable/Checklist.jsx
@@ -1,22 +1,5 @@
 import * as Checklist from '../Checklist'
-import themeable from '../lib/decorators/themeable'
 
-export const Main = themeable(Checklist.Main, (customizations, props) => ({
-  customize: {
-    ...props.customize,
-    borderColor: customizations.color_border,
-    borderRadius: customizations.radius_border
-  }
-}))
+export const Main = Checklist.Main
 
-Main.displayName = 'ThemeableChecklist.Main'
-
-export const Item = themeable(Checklist.Item, (customizations, props) => ({
-  customize: {
-    ...props.customize,
-    strokeColor: customizations.color_details,
-    textColor: customizations.color_text
-  }
-}))
-
-Item.displayName = 'ThemeableChecklist.Item'
+export const Item = Checklist.Item
