+++ themeable/Link.jsx
@@ -1,5 +1,5 @@
 import Link from '../Link'
-import themeable from '../lib/decorators/themeable'
+import themeable from '../decorators/themeable'
 
 const Component = themeable(Link, (customizations, props) => ({
   customize: {
